size_t case

+++ 0x1E-search_algorithms/0-linear.c
@@ -12,7 +12,7 @@
 
 int linear_search(int *array, size_t size, int value)
 {
-	size_t int i;
+	unsigned int i;
 
 	if (array != NULL)
 	{
